Increase the page size of the fetch GH releases command. (#79)
We were in a situation where we haven't released a new version for a while and created a lot of app store release candidates since the last release.
As the GH api releases endpoint is paginated with a default page size of 30 the response didn't contain the latest no release tag. If we would support pagination it would be returned on the page 2.
The easiest fix at this time is to simply set the page size to the max (100). It is unlikely that within our current release process we will exceed this number.